RSFB-2638: changes the return of div0 to decimal

+++ core/src/main/java/org/apache/calcite/sql/fun/SqlLibraryOperators.java
@@ -2061,7 +2061,7 @@ public abstract class SqlLibraryOperators {
   public static final SqlFunction DIV0 =
       new SqlFunction("DIV0",
           SqlKind.OTHER_FUNCTION,
-          ReturnTypes.INTEGER, null,
+          ReturnTypes.DECIMAL, null,
           OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.NUMERIC),
           SqlFunctionCategory.NUMERIC);
 }
